+++ BYOJoystick/Managers/AH6Manager.cs
@@ -44,6 +44,7 @@ namespace BYOJoystick.Managers
 
         protected override void PreMapping()
         {
+            LogInteractablesIfEnabled(ShortName);
         }
 
         protected override void CreateFlightControls()