+++ BYOJoystick/Managers/AH94RearManager.cs
@@ -28,6 +28,7 @@ namespace BYOJoystick.Managers
 
         protected override void PreMapping()
         {
+            LogInteractablesIfEnabled(ShortName);
         }
 
         protected override void CreateFlightControls()