+++ BYOJoystick/Managers/AV42CManager.cs
@@ -22,6 +22,7 @@ namespace BYOJoystick.Managers
 
         protected override void PreMapping()
         {
+            LogInteractablesIfEnabled(ShortName);
         }
 
         protected override void CreateFlightControls()