+++ BYOJoystick/Managers/T55RearManager.cs
@@ -27,6 +27,7 @@ namespace BYOJoystick.Managers
 
         protected override void PreMapping()
         {
+            LogInteractablesIfEnabled(ShortName);
         }
 
         protected override void CreateFlightControls()